Allow ruby scripts to run in WSL
I didn't want to install Ruby on my Windows machine, but I already had
it installed in my Ubuntu distro in my Windows Subsystem for Linux. The
only thing is WSL1 doesn't support the socket option SO_LINGER. So this
commit detects wether or not the tests are running in a WSL environment
and if detected, does not attempt to set the SO_LINGER option.