Removes setup.bat; Adds WSL support
This commit removes the setup.bat file and instead adds that
functionality into the runat.ps1 PowerShell script, which near as far as
I can tell, was the only place setup.bat was used.

This simplifies the web of scripts used to run the acceptance tests as
well as centralizes on using PowerShell throughout instead of a mix of
batch scripts and PowerShell scripts.

The runat.ps1 PowerShell script was reformatted to follow
PowerShell formatting idioms. The usage function was removed--instead,
the parameters were updated with PowerShell cmdlet parameter metadata.
You can find "usage" by using `Get-Help .\runat.ps1`.

The KillChildren function underwent a couple of changes. I'm not sure if
something was wrong with it before, but it happened to work "enough" to
allow the tests to be run. But I found locally, it was killing my
PowerShell session, terminating the acceptance tests after only the
first batch. Then, up on the CI server, the test run would hang. As far
as I can tell, for the local termination problem, it seems to be linked
to the killing of the child process conhost.exe, which is the console
host for terminal apps on Windows. So KillChildren now filters the child
processes to exclude conhost.exe. Furthermore, in Get-WmiObject is
obsolete. And in PowerShell Core, it doesn't exist at all. So it has
been replaced with Get-CimInstance, which is the recommended
replacement.

Lastly, instead of just dumping the test results XML to the console, I
leveraged PowerShell to get the contents of the file and display a table
of the tests and whether they passed, as well as the overall result
status.